chg: [example] `followed_by` array added
In a series of inject, you want to describe the order of injection
which can be loosy tight to a `requirements` or not. `followed_by` removes
the ambiguity of the flow order and can be loosly used in simple
scenario without the need to have any specific requirements or step.